SERVER - Refactor Category and Order

+++ server/src/main/java/com/shoppingsystem/shopping_system/category/service/CategoryServiceImpl.java
@@ -1,6 +1,7 @@
 package com.shoppingsystem.shopping_system.category.service;
 
-import com.shoppingsystem.shopping_system.category.dto.CategoryDTO;
+import com.shoppingsystem.shopping_system.category.dto.CategoryResponse;
+import com.shoppingsystem.shopping_system.category.exceptions.CategoryNotFoundException;
 import com.shoppingsystem.shopping_system.category.model.Category;
 import com.shoppingsystem.shopping_system.category.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +9,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -22,36 +22,46 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public List<CategoryDTO> findAll() {
+    public List<CategoryResponse> findAll() {
         List<Category> categories = categoryRepository.findAll();
+        if(categories.isEmpty()) {
+            throw new CategoryNotFoundException("No categories were found");
+        }
         return categories.stream().map(this::convertToDTO).collect(Collectors.toList());
     }
 
     @Override
-    public Category findById(int id) {
-        Optional<Category> result = categoryRepository.findById(id);
-        Category theCategory = null;
-        if(result.isPresent()) {
-            theCategory = result.get();
-        } else {
-            throw new RuntimeException("Did not find product with id - " + id);
-        }
-        return theCategory;
+    public CategoryResponse findById(int id) {
+        return categoryRepository.findById(id)
+                .map(this::convertToDTO)
+                .orElseThrow(() -> new CategoryNotFoundException("Category not found with id - " + id ));
+    }
+
+    @Override
+    public Category findByIdEntity(int id) {
+        return categoryRepository.findById(id).orElseThrow(() -> new CategoryNotFoundException("Couldn't found category" +
+                "with ID - " + id));
     }
 
     @Override
     @Transactional
     public Category save(Category category) {
-        return (Category) categoryRepository.save(category);
+        if(category == null || category.getName() == null || category.getName().trim().isEmpty()) {
+            throw new IllegalArgumentException("Category must not be null");
+        }
+        return categoryRepository.save(category);
     }
 
     @Override
     @Transactional
     public void deleteById(int id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new CategoryNotFoundException("Category not found with id - " + id);
+        }
         categoryRepository.deleteById(id);
     }
 
-    public CategoryDTO convertToDTO(Category category) {
-        return new CategoryDTO(category.getId(), category.getName());
+    public CategoryResponse convertToDTO(Category category) {
+        return new CategoryResponse(category.getId(), category.getName());
     }
 }
